code formatting improved

+++ src/components/FormContainer.js
@@ -21,15 +21,15 @@ class FormContainer extends Component {
         }
         this.promise = db.structures.toArray().then(response => {
             return response;
-            }).then(state => {
-                const flatState = this.transformToFlatStructure(state);
-                this.setState({structure: flatState});
+        }).then(state => {
+            const flatState = this.transformToFlatStructure(state);
+            this.setState({ structure: flatState });
         })
     }
 
     transformToFlatStructure(array) {
         var result = [];
-        array.forEach( a => {
+        array.forEach(a => {
             result.push(a);
             if (Array.isArray(a.children)) {
                 result = result.concat(this.transformToFlatStructure(a.children));
@@ -38,14 +38,14 @@ class FormContainer extends Component {
         return result;
     };
 
-    addDepth(arr, depth = 0){
+    addDepth(arr, depth = 0) {
         arr.forEach(obj => {
             obj.depth = depth
-            this.addDepth(obj.children, depth+1)
+            this.addDepth(obj.children, depth + 1)
         })
     }
 
-    transformToTree(array){
+    transformToTree(array) {
         let map = {}, node, result = [];
         for (let i = 0; i < array.length; i++) {
             map[array[i].id] = i;
@@ -53,10 +53,10 @@ class FormContainer extends Component {
         }
 
         for (let i = 0; i < array.length; i++) {
-            
+
             node = array[i];
             if (node.parentId !== '0') {
-                if (array[map[node.parentId]]){
+                if (array[map[node.parentId]]) {
                     array[map[node.parentId]].children.push(node);
                 }
             } else {
@@ -69,50 +69,50 @@ class FormContainer extends Component {
     generateNewId() {
         let idsArray = [];
 
-        for (let i = 0; i < this.state.structure.length; i++){
-           idsArray.push(parseInt(this.state.structure[i].id));
+        for (let i = 0; i < this.state.structure.length; i++) {
+            idsArray.push(parseInt(this.state.structure[i].id));
         }
         return (Math.max(...idsArray) + 1).toLocaleString();
     };
 
-    updateDB(){
+    updateDB() {
         const n = 0;
-        Dexie.spawn(function*() {
+        Dexie.spawn(function* () {
             yield db.structures
-            .where('id')
-            .above(n.toLocaleString())
-            .delete();
-        }).catch (err => {
-            console.error ('Deleting from db failed' + err.stack);
+                .where('id')
+                .above(n.toLocaleString())
+                .delete();
+        }).catch(err => {
+            console.error('Deleting from db failed' + err.stack);
         });
-        let  inputs = this.transformToTree(this.state.structure);    
+        let inputs = this.transformToTree(this.state.structure);
         this.addDepth(inputs);
 
-        Dexie.spawn(function*() {
+        Dexie.spawn(function* () {
             yield db.structures.bulkPut(inputs);
-            }).catch (err => {
-            console.error ('Ooops' + err.stack);
-            });
+        }).catch(err => {
+            console.error('Ooops' + err.stack);
+        });
     };
-    
-    handleAddMainInput(e){
+
+    handleAddMainInput(e) {
         e.preventDefault();
         let inputs = this.state.structure;
         let newInput;
 
         if (inputs.length > 0) {
             newInput = {
-                    'id': this.generateNewId(),
-                    'parentId': '0', 
-                    'type': 'main'
-                }
+                'id': this.generateNewId(),
+                'parentId': '0',
+                'type': 'main'
+            }
         } else newInput = {
             'id': '1',
-            'parentId': '0', 
+            'parentId': '0',
             'type': 'main'
         };
         inputs.push(newInput);
-        this.setState({structure: inputs}, this.updateDB);    
+        this.setState({ structure: inputs }, this.updateDB);
     };
 
     handleAddSubInput(i, question, type, firstConditionField, secondConditionField) {
@@ -129,20 +129,21 @@ class FormContainer extends Component {
 
         newStructure[i].question = question;
         newStructure[i].conditionType = type;
-        if(firstConditionField){
+        if (firstConditionField) {
             newStructure[i].firstConditionField = firstConditionField;
         }
-        if (secondConditionField){
+        if (secondConditionField) {
             newStructure[i].secondConditionField = secondConditionField;
         }
         this.setState({
-            structure: newStructure, 
+            structure: newStructure,
             newInput: {
-                text: question, 
+                text: question,
                 select: type,
                 firstConditionFieldValue: firstConditionField,
                 secondConditionFieldValue: secondConditionField
-            }}, this.updateDB);  
+            }
+        }, this.updateDB);
     };
 
     handleDeleteInput(i) {
@@ -153,46 +154,46 @@ class FormContainer extends Component {
         inputs.splice(i, 1);
 
         for (let i = start; i < inputs.length; i++) {
-            if (inputs[i]){
-                if (inputs[i].parentId === inputId){
+            if (inputs[i]) {
+                if (inputs[i].parentId === inputId) {
                     inputId = inputs[i].id;
                     inputs.splice(i, 1);
-                    i--; 
+                    i--;
                 }
             }
         }
         let mainInputs = 0;
         inputs.forEach(e => {
-            if(e.type === 'main') mainInputs++;
+            if (e.type === 'main') mainInputs++;
         })
         if (mainInputs === 0) inputs = [];
 
-        this.setState({structure: inputs}, this.updateDB);
+        this.setState({ structure: inputs }, this.updateDB);
     };
 
     render() {
         const inputs = this.state.structure.map((e, i) => {
-            const margin = (e.depth+1)*30+'px'
+            const margin = (e.depth + 1) * 30 + 'px'
             if (e.type === 'sub') return (
-                <SubInput type='Number' 
-                          handleAddSubInput={(e, question, type, firstConditionField, secondConditionField) => this.handleAddSubInput(e, question, type, firstConditionField, secondConditionField)} 
-                          handleDelete={e => this.handleDeleteInput(e)} 
-                          input={this.state.newInput}
-                          index={i} 
-                          margin={margin}
-                          key={i}/>
+                <SubInput type='Number'
+                    handleAddSubInput={(e, question, type, firstConditionField, secondConditionField) => this.handleAddSubInput(e, question, type, firstConditionField, secondConditionField)}
+                    handleDelete={e => this.handleDeleteInput(e)}
+                    input={this.state.newInput}
+                    index={i}
+                    margin={margin}
+                    key={i} />
             )
             else return (
-                <MainInput handleAddSubInput={(e, question, type) => this.handleAddSubInput(e, question, type)} 
-                           handleDelete={e => this.handleDeleteInput(e)} 
-                           index={i} 
-                           key ={i}/>
+                <MainInput handleAddSubInput={(e, question, type) => this.handleAddSubInput(e, question, type)}
+                    handleDelete={e => this.handleDeleteInput(e)}
+                    index={i}
+                    key={i} />
             )
         })
         return (
             <form className='form-group'>
                 {inputs}
-                <button className='btn'onClick={e => this.handleAddMainInput(e)}>Add Input</button>
+                <button className='btn' onClick={e => this.handleAddMainInput(e)}>Add Input</button>
             </form>
         )
     }
